+++ Example/StripeTerminal.xcframework/ios-arm64_x86_64-simulator/StripeTerminal.framework/Headers/SCPErrors.h
@@ -165,6 +165,7 @@ typedef NS_ERROR_ENUM(SCPErrorDomain, SCPError){
      */
     SCPErrorInvalidLocationIdParameter = 1960,
 
+
     /*
      USER ERRORS
      */